bug fix su casting di tipo

+++ src/main/java/it/corso/Main.java
@@ -4,10 +4,10 @@ public class Main {
     public static void main(String[] args) {
 
         LibraryService libreria = new LibraryService();
-        libreria.addBook( new Book(1, "titolo1", "Nicolò") );
-        libreria.addBook( new Book(2, "titolo2", "Giacomo") );
-        libreria.addBook( new Book(3, "titolo3", "Edoardo") );
-        libreria.addBook( new Book(4, "titolo4", "Giovanna") );
+        libreria.addBook( new Book((long)1, "titolo1", "Nicolò") );
+        libreria.addBook( new Book((long)2, "titolo2", "Giacomo") );
+        libreria.addBook( new Book((long)3, "titolo3", "Edoardo") );
+        libreria.addBook( new Book((long)4, "titolo4", "Giovanna") );
 
         libreria.listBooks();
